feat(native): add native sampler, Perfetto export and fairer ring-buffer eviction
- NativeSampler thread records process CPU/memory and UI frame windows
  (fps, max frame gap, slow and frozen frames) with source=native; frames
  come from CADisplayLink on iOS and Choreographer on Android via JNI.
- Recording.exportTraceEvents() emits Chrome Trace Event / Perfetto JSON
  with strictly nested lanes per source, instants and counters.
- Metric samples may fill at most half of maxEvents and replace older
  samples, so periodic sampling no longer evicts spans and marks in long
  sessions (a 95 minute live session kept only ~5 minutes of spans).
- Cast range-checked limits to size_t so armeabi-v7a builds compile.
- Ship consumer ProGuard rules for the frame source.

+++ packages/react-native-nitro-tracing/cpp/NativeConversions.cpp
@@ -15,8 +15,10 @@ tracingcore::Context toCore(const std::string &name, const std::string &id, cons
   return c;
 }
 tracingcore::Config toCore(const RecordingOptions &o) {
-  return {checkedInteger(o.maxEvents, 100000, "maxEvents"), checkedInteger(o.maxBytes, 67108864, "maxBytes"),
-          checkedInteger(o.maxActiveSpans, 10000, "maxActiveSpans"), o.spanTimeoutMs};
+  // Range-checked above, so the casts are safe where size_t is 32-bit (armeabi-v7a).
+  return {static_cast<size_t>(checkedInteger(o.maxEvents, 100000, "maxEvents")),
+          static_cast<size_t>(checkedInteger(o.maxBytes, 67108864, "maxBytes")),
+          static_cast<size_t>(checkedInteger(o.maxActiveSpans, 10000, "maxActiveSpans")), o.spanTimeoutMs};
 }
 tracingcore::Outcome toCore(SpanOutcome o) {
   switch (o) {
